refactor: split internal API HTTP routes
