+++ internal-api/src/http_router.rs
@@ -1,283 +1,24 @@
 use axum::{
-    extract::{Path, Query, Request},
+    extract::{Path, Query},
     http::{HeaderMap, Method, StatusCode},
-    middleware::{self, Next},
-    response::{IntoResponse, Json, Response},
+    middleware,
+    response::{IntoResponse, Json},
     routing::{get, post, put},
     Router,
 };
-use log::error;
 use serde::Deserialize;
 use serde_json::{json, Value};
 use std::collections::HashMap;
 use tower_http::cors::{Any, CorsLayer};
 
-use env_common::errors::ModuleError;
-
+use crate::deployment_routes::run_claim;
 use crate::handlers;
-
-fn status_code_for_module_error(e: &ModuleError) -> StatusCode {
-    match e {
-        ModuleError::ModuleVersionExists(_, _) => StatusCode::CONFLICT,
-        ModuleError::ValidationError(_) => StatusCode::CONFLICT,
-        ModuleError::InvalidTrack(_)
-        | ModuleError::InvalidTrackPrereleaseVersion(_, _)
-        | ModuleError::InvalidStableVersion
-        | ModuleError::InvalidModuleSchema(_)
-        | ModuleError::InvalidExampleVariable(_)
-        | ModuleError::InvalidVariableNaming(_)
-        | ModuleError::InvalidOutputNaming(_)
-        | ModuleError::InvalidReference(_, _)
-        | ModuleError::ModuleVersionNotSet(_)
-        | ModuleError::ModuleVersionMissing(_)
-        | ModuleError::DuplicateClaimNames(_)
-        | ModuleError::CircularDependency(_)
-        | ModuleError::SelfReferencingClaim(_, _, _)
-        | ModuleError::StackModuleNamespaceIsSet(_)
-        | ModuleError::TerraformLockfileExists()
-        | ModuleError::TerraformLockfileEmpty
-        | ModuleError::TerraformNoLockfile(_)
-        | ModuleError::NoProvidersDefined(_)
-        | ModuleError::NoRequiredProvidersDefined(_)
-        | ModuleError::OutputKeyNotFound(_, _, _, _, _)
-        | ModuleError::StackClaimReferenceNotFound(_, _, _, _)
-        | ModuleError::UnresolvedReference(_, _) => StatusCode::BAD_REQUEST,
-        ModuleError::ModuleVersionNotFound(_, _) => StatusCode::NOT_FOUND,
-        ModuleError::UploadModuleError(_)
-        | ModuleError::ZipError(_)
-        | ModuleError::PublishError(_)
-        | ModuleError::Other(_) => StatusCode::INTERNAL_SERVER_ERROR,
-    }
-}
-
-// Helper function to handle responses consistently
-async fn handle_result(result: anyhow::Result<Value>) -> impl IntoResponse {
-    match result {
-        Ok(mut response) => {
-            // Check if response is an object with "Items"
-            if let Some(obj) = response.as_object_mut() {
-                if obj.contains_key("Items") {
-                    // It's a list response
-                    let items = obj.remove("Items").unwrap();
-
-                    // Check for pagination token
-                    let mut headers = axum::http::HeaderMap::new();
-
-                    // Handle "next_token" if it exists
-                    if let Some(next_token) = obj.remove("next_token") {
-                        if let Some(token_str) = next_token.as_str() {
-                            if let Ok(val) = axum::http::HeaderValue::from_str(token_str) {
-                                headers.insert("x-next-token", val);
-                            }
-                        }
-                    }
-
-                    return (StatusCode::OK, headers, Json(items)).into_response();
-                }
-            }
-
-            // Default behavior for non-list responses
-            (StatusCode::OK, Json(response)).into_response()
-        }
-        Err(e) => {
-            let err_msg = e.to_string();
-
-            // Try to downcast to ModuleError for typed status code mapping
-            let status = if let Some(module_err) = e.downcast_ref::<ModuleError>() {
-                status_code_for_module_error(module_err)
-            } else if err_msg.to_lowercase().contains("not found") {
-                StatusCode::NOT_FOUND
-            } else {
-                StatusCode::INTERNAL_SERVER_ERROR
-            };
-
-            if status == StatusCode::INTERNAL_SERVER_ERROR {
-                error!("Request failed: {:?}", e);
-            }
-
-            // Include full error details (including cause chain) for internal server errors
-            // This aids debugging client-side when using the API directly
-            let response_msg = if status == StatusCode::INTERNAL_SERVER_ERROR {
-                format!("{:?}", e)
-            } else {
-                err_msg
-            };
-
-            (
-                status,
-                Json(json!({
-                    "error": response_msg
-                })),
-            )
-                .into_response()
-        }
-    }
-}
-
-async fn auth_middleware(
-    headers: HeaderMap,
-    Path(params): Path<HashMap<String, String>>,
-    request: Request,
-    next: Next,
-) -> Response {
-    if let Err(e) = ensure_authenticated_user(&headers) {
-        return e.into_response();
-    }
-
-    if let Some(project_param) = params.get("project") {
-        for project in project_param.split(',') {
-            let p = project.trim();
-            if !p.is_empty() {
-                if let Err(e) = ensure_access(&headers, p).await {
-                    return e.into_response();
-                }
-            }
-        }
-    }
-
-    next.run(request).await
-}
-
-/// Middleware that enforces publish permissions with the Rego publish policy.
-///
-/// Extracts the resource type from the URL path and the resource name from
-/// either path parameters (for deprecate) or the request body (for publish).
-async fn publish_auth_middleware(headers: HeaderMap, request: Request, next: Next) -> Response {
-    let path = request.uri().path().to_string();
-    let method = request.method().clone();
-
-    // Determine resource type and name from the request
-    let (resource_type, resource_name, resource_track) =
-        if method == Method::PUT && path.contains("/deprecate") {
-            // Deprecate routes:
-            //   /api/v1/module/:track/:module/:version/deprecate
-            //   /api/v1/stack/:track/:stack/:version/deprecate
-            let segments: Vec<&str> = path.split('/').collect();
-            // segments: ["", "api", "v1", "module"|"stack", track, name, version, "deprecate"]
-            if segments.len() >= 6 {
-                let res_type = if segments[3] == "stack" {
-                    "stack"
-                } else {
-                    "module"
-                };
-                (
-                    res_type.to_string(),
-                    segments[5].to_string(),
-                    Some(segments[4].to_string()),
-                )
-            } else {
-                return (
-                    StatusCode::BAD_REQUEST,
-                    Json(json!({ "error": "Invalid deprecate path" })),
-                )
-                    .into_response();
-            }
-        } else if method == Method::POST {
-            // Publish routes: read body to extract resource name
-            let (parts, body) = request.into_parts();
-            let bytes = match axum::body::to_bytes(body, 512 * 1024 * 1024).await {
-                Ok(b) => b,
-                Err(e) => {
-                    return (
-                        StatusCode::BAD_REQUEST,
-                        Json(json!({ "error": format!("Failed to read request body: {}", e) })),
-                    )
-                        .into_response();
-                }
-            };
-
-            let body_json: Value = match serde_json::from_slice(&bytes) {
-                Ok(v) => v,
-                Err(e) => {
-                    return (
-                        StatusCode::BAD_REQUEST,
-                        Json(json!({ "error": format!("Invalid JSON body: {}", e) })),
-                    )
-                        .into_response();
-                }
-            };
-
-            let (res_type, res_name) = if path.contains("/module/publish") {
-                let name = body_json
-                    .get("module")
-                    .and_then(|m| m.get("module").or_else(|| m.get("module_name")))
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("unknown")
-                    .to_string();
-                ("module".to_string(), name)
-            } else if path.contains("/stack/publish") {
-                let name = body_json
-                    .get("module")
-                    .and_then(|m| m.get("module").or_else(|| m.get("module_name")))
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("unknown")
-                    .to_string();
-                ("stack".to_string(), name)
-            } else if path.contains("/provider/publish") {
-                let name = body_json
-                    .get("provider")
-                    .and_then(|p| p.get("name"))
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("unknown")
-                    .to_string();
-                ("provider".to_string(), name)
-            } else if path.contains("/policy/publish") {
-                let name = body_json
-                    .get("policy")
-                    .and_then(|p| p.get("policy"))
-                    .and_then(|v| v.as_str())
-                    .unwrap_or("unknown")
-                    .to_string();
-                ("policy".to_string(), name)
-            } else {
-                return (
-                    StatusCode::BAD_REQUEST,
-                    Json(json!({ "error": "Unknown publish endpoint" })),
-                )
-                    .into_response();
-            };
-            let track = body_json
-                .get("track")
-                .and_then(|v| v.as_str())
-                .map(str::to_string)
-                .or_else(|| {
-                    body_json
-                        .get("module")
-                        .and_then(|m| m.get("track"))
-                        .and_then(|v| v.as_str())
-                        .map(str::to_string)
-                });
-
-            // Reconstruct the request with the buffered body
-            let request = Request::from_parts(parts, axum::body::Body::from(bytes));
-            // Check permissions before continuing
-            if let Err(e) =
-                ensure_publish_access(&headers, &res_type, &res_name, track.as_deref()).await
-            {
-                return e.into_response();
-            }
-            return next.run(request).await;
-        } else {
-            return (
-                StatusCode::METHOD_NOT_ALLOWED,
-                Json(json!({ "error": "Unsupported method for publish endpoint" })),
-            )
-                .into_response();
-        };
-
-    // Check permissions (for non-POST paths like deprecate)
-    if let Err(e) = ensure_publish_access(
-        &headers,
-        &resource_type,
-        &resource_name,
-        resource_track.as_deref(),
-    )
-    .await
-    {
-        return e.into_response();
-    }
-    next.run(request).await
-}
+use crate::http_authz::{auth_middleware, extract_jwt_claims, publish_auth_middleware};
+use crate::http_response::handle_result;
+use crate::publish_routes::{
+    deprecate_module, deprecate_stack, download_provider, publish_module, publish_policy,
+    publish_provider, publish_stack,
+};
 
 pub fn create_router() -> Router {
     // Configure CORS to allow requests from any origin
@@ -431,223 +172,6 @@ pub fn create_router() -> Router {
     // NOTE: CompressionLayer removed because API Gateway v2 HTTP API strips the
     // Content-Encoding header, causing clients to receive compressed data without
     // knowing it's compressed. Use CloudFront for compression instead.
-}
-
-/// Extract and decode JWT token from Authorization header without signature validation.
-/// We only read the claims without verifying the signature; the signature is validated
-/// upstream by the platform (API Gateway on AWS, or Azure App Service EasyAuth on Azure)
-/// before the request ever reaches this service.
-fn extract_jwt_claims(headers: &HeaderMap) -> Option<Value> {
-    if let Some(claims) = extract_verified_claims(headers) {
-        return Some(claims);
-    }
-
-    let auth_header = headers.get("authorization").and_then(|v| v.to_str().ok())?;
-
-    // Remove "Bearer " prefix
-    let token = auth_header
-        .strip_prefix("Bearer ")
-        .or_else(|| auth_header.strip_prefix("bearer "))?;
-
-    // JWT format: header.payload.signature
-    let parts: Vec<&str> = token.split('.').collect();
-    if parts.len() != 3 {
-        log::warn!("Invalid JWT format: expected 3 parts, got {}", parts.len());
-        return None;
-    }
-
-    // Decode the payload (second part) without verification
-    use base64::{engine::general_purpose::URL_SAFE_NO_PAD, Engine as _};
-    let payload_bytes = match URL_SAFE_NO_PAD.decode(parts[1]) {
-        Ok(bytes) => bytes,
-        Err(e) => {
-            log::warn!("Failed to decode JWT payload: {}", e);
-            return None;
-        }
-    };
-
-    match serde_json::from_slice(&payload_bytes) {
-        Ok(claims) => Some(claims),
-        Err(e) => {
-            log::error!("Failed to parse JWT claims as JSON: {}", e);
-            None
-        }
-    }
-}
-
-fn extract_verified_claims(headers: &HeaderMap) -> Option<Value> {
-    let claims_header = headers
-        .get("x-infraweave-verified-claims")
-        .and_then(|v| v.to_str().ok())?;
-
-    use base64::{engine::general_purpose::URL_SAFE_NO_PAD, Engine as _};
-
-    let payload_bytes = match URL_SAFE_NO_PAD.decode(claims_header) {
-        Ok(bytes) => bytes,
-        Err(e) => {
-            log::warn!(
-                "Failed to decode x-infraweave-verified-claims header: {}",
-                e
-            );
-            return None;
-        }
-    };
-
-    match serde_json::from_slice(&payload_bytes) {
-        Ok(claims) => Some(claims),
-        Err(e) => {
-            log::error!(
-                "Failed to parse x-infraweave-verified-claims as JSON: {}",
-                e
-            );
-            None
-        }
-    }
-}
-
-async fn ensure_access(
-    headers: &HeaderMap,
-    project_id: &str,
-) -> Result<(), (StatusCode, axum::response::Json<serde_json::Value>)> {
-    if let Some(_user_id) = headers.get("x-auth-user").and_then(|v| v.to_str().ok()) {
-        // Extract JWT claims from Authorization header
-        if let Some(claims) = extract_jwt_claims(headers) {
-            // Check for allowed_projects claim (configurable via AUTH_ALLOWED_PROJECTS_CLAIM)
-            let claim_key = crate::auth_handler::allowed_projects_claim_key();
-            if let Some(allowed_projects_str) = claims.get(&claim_key).and_then(|v| v.as_str()) {
-                let allowed_projects: Vec<String> = allowed_projects_str
-                    .split(',')
-                    .map(|s| s.trim().to_string())
-                    .filter(|s| !s.is_empty())
-                    .collect();
-
-                if allowed_projects.contains(&project_id.to_string()) {
-                    return Ok(());
-                } else {
-                    return Err((
-                        StatusCode::FORBIDDEN,
-                        Json(json!({
-                            "error": "Access denied to this project"
-                        })),
-                    ));
-                }
-            }
-        }
-
-        // No allowed_projects claim found in JWT; deny access.
-        log::warn!(
-            "User has no '{}' claim in JWT; denying access to project",
-            crate::auth_handler::allowed_projects_claim_key(),
-        );
-        Err((
-            StatusCode::FORBIDDEN,
-            Json(json!({
-                "error": "Access denied: no allowed_projects claim found in token"
-            })),
-        ))
-    } else {
-        #[cfg(feature = "local")]
-        {
-            log::warn!(
-                "Missing x-auth-user header, allowing access to project {} (LOCAL MODE ONLY)",
-                project_id
-            );
-            Ok(())
-        }
-        #[cfg(not(feature = "local"))]
-        {
-            Err((
-                StatusCode::UNAUTHORIZED,
-                Json(json!({
-                    "error": "Missing authentication user context"
-                })),
-            ))
-        }
-    }
-}
-
-fn ensure_authenticated_user(
-    headers: &HeaderMap,
-) -> Result<(), (StatusCode, axum::response::Json<serde_json::Value>)> {
-    if headers
-        .get("x-auth-user")
-        .and_then(|v| v.to_str().ok())
-        .is_some()
-    {
-        return Ok(());
-    }
-
-    #[cfg(feature = "local")]
-    {
-        log::warn!(
-            "Missing x-auth-user header, allowing authenticated-only route (LOCAL MODE ONLY)"
-        );
-        Ok(())
-    }
-    #[cfg(not(feature = "local"))]
-    {
-        Err((
-            StatusCode::UNAUTHORIZED,
-            Json(json!({
-                "error": "Missing authentication user context"
-            })),
-        ))
-    }
-}
-
-/// Ensure the authenticated user has permission to publish a specific resource.
-///
-/// Authorization is handled by the configured identity-provider extractor and
-/// Rego publish policy (see [`crate::publish_auth`]).
-///
-/// For CI/CD pipelines: configure the pipeline's identity in your identity provider
-/// so the Rego policy can derive the publish principal from trusted claims.
-async fn ensure_publish_access(
-    headers: &HeaderMap,
-    resource_type: &str,
-    resource_name: &str,
-    track: Option<&str>,
-) -> Result<(), (StatusCode, axum::response::Json<serde_json::Value>)> {
-    let resource_desc = match track {
-        Some(track) => format!("{}/{}/{}", resource_type, resource_name, track),
-        None => format!("{}/{}", resource_type, resource_name),
-    };
-
-    if let Some(claims) = extract_verified_claims(headers) {
-        if crate::publish_auth::check_publish(&claims, resource_type, resource_name, track).await {
-            log::info!(
-                "User authorized to publish {} via Rego policy",
-                resource_desc
-            );
-            return Ok(());
-        }
-
-        log::warn!("User has no matching publish grant for {}", resource_desc);
-        Err((
-            StatusCode::FORBIDDEN,
-            Json(json!({
-                "error": format!("You do not have permission to publish {}. Configure AUTH_PUBLISH_REGO_POLICY_PARAMETER with the Rego publish policy.", resource_desc)
-            })),
-        ))
-    } else {
-        #[cfg(feature = "local")]
-        {
-            log::warn!(
-                "Missing x-auth-user header, allowing publish access to {} (LOCAL MODE ONLY)",
-                resource_desc
-            );
-            Ok(())
-        }
-        #[cfg(not(feature = "local"))]
-        {
-            Err((
-                StatusCode::UNAUTHORIZED,
-                Json(json!({
-                    "error": "Missing authentication user context"
-                })),
-            ))
-        }
-    }
 }
 
 // Handler implementations
@@ -1413,222 +937,6 @@ async fn get_provider_download_url(
     }
 }
 
-#[derive(Deserialize)]
-struct DeprecateModuleBody {
-    message: Option<String>,
-}
-
-async fn deprecate_module(
-    Path((track, module, version)): Path<(String, String, String)>,
-    Json(body): Json<DeprecateModuleBody>,
-) -> impl IntoResponse {
-    handle_result(
-        handlers::deprecate_module(&json!({
-            "track": track,
-            "module": module,
-            "version": version,
-            "message": body.message
-        }))
-        .await,
-    )
-    .await
-}
-
-async fn deprecate_stack(
-    Path((track, stack, version)): Path<(String, String, String)>,
-    Json(body): Json<DeprecateModuleBody>,
-) -> impl IntoResponse {
-    handle_result(
-        handlers::deprecate_stack(&json!({
-            "track": track,
-            "stack": stack,
-            "version": version,
-            "message": body.message
-        }))
-        .await,
-    )
-    .await
-}
-
-#[derive(Deserialize)]
-struct PublishModuleBody {
-    zip_base64: String,
-    module: Value, // ModuleResp serialized as JSON
-    track: String,
-    version: String,
-    job_id: String,
-}
-
-async fn download_provider(Json(body): Json<Value>) -> impl IntoResponse {
-    handle_result(
-        handlers::download_provider(&json!({
-            "s3_key": body.get("s3_key")
-        }))
-        .await,
-    )
-    .await
-    .into_response()
-}
-
-async fn publish_module(Json(body): Json<PublishModuleBody>) -> impl IntoResponse {
-    handle_result(
-        handlers::publish_module(&json!({
-            "zip_base64": body.zip_base64,
-            "module": body.module,
-            "track": body.track,
-            "version": body.version,
-            "job_id": body.job_id
-        }))
-        .await,
-    )
-    .await
-}
-
-/// Stack publish handler - stacks use the same ModuleResp format as modules,
-/// but the server-side handler also ensures provider caches exist in all regions.
-async fn publish_stack(Json(body): Json<PublishModuleBody>) -> impl IntoResponse {
-    handle_result(
-        handlers::publish_stack(&json!({
-            "zip_base64": body.zip_base64,
-            "module": body.module,
-            "track": body.track,
-            "version": body.version,
-            "job_id": body.job_id
-        }))
-        .await,
-    )
-    .await
-}
-
-#[derive(Deserialize)]
-struct PublishProviderBody {
-    zip_base64: String,
-    provider: Value,
-}
-
-#[derive(Deserialize)]
-struct PublishPolicyBody {
-    zip_base64: String,
-    policy: Value,
-}
-
-async fn publish_provider(Json(body): Json<PublishProviderBody>) -> impl IntoResponse {
-    handle_result(
-        handlers::publish_provider(&json!({
-            "zip_base64": body.zip_base64,
-            "provider": body.provider
-        }))
-        .await,
-    )
-    .await
-}
-
-async fn publish_policy(Json(body): Json<PublishPolicyBody>) -> impl IntoResponse {
-    handle_result(
-        handlers::publish_policy(&json!({
-            "zip_base64": body.zip_base64,
-            "policy": body.policy
-        }))
-        .await,
-    )
-    .await
-}
-
-async fn run_claim(headers: HeaderMap, Json(body): Json<Value>) -> Response {
-    log::info!("Received run_claim request");
-    // Body is ApiInfraPayloadWithVariables
-    // Manually extract payload and variables from the JSON value
-    let payload_value = match body.get("payload") {
-        Some(p) => p.clone(),
-        None => {
-            return handle_result(Err(anyhow::anyhow!("Missing 'payload' field")))
-                .await
-                .into_response()
-        }
-    };
-
-    let variables = match body.get("variables") {
-        Some(v) => v.clone(),
-        None => {
-            return handle_result(Err(anyhow::anyhow!("Missing 'variables' field")))
-                .await
-                .into_response()
-        }
-    };
-
-    let payload: env_defs::ApiInfraPayload = match serde_json::from_value(payload_value.clone()) {
-        Ok(p) => p,
-        Err(e) => {
-            return handle_result(Err(anyhow::anyhow!("Invalid payload: {}", e)))
-                .await
-                .into_response()
-        }
-    };
-
-    if let Err(e) = ensure_access(&headers, &payload.project_id).await {
-        return e.into_response();
-    }
-
-    handlers::with_workload_account(
-        payload.project_id.clone(),
-        run_claim_authorized(payload_value, variables, payload),
-    )
-    .await
-}
-
-async fn run_claim_authorized(
-    payload_value: Value,
-    variables: Value,
-    payload: env_defs::ApiInfraPayload,
-) -> Response {
-    // Launch runner with ApiInfraPayload only (no variables to avoid size limits)
-    let result = handlers::start_runner(&json!({
-        "data": payload_value
-    }))
-    .await;
-
-    let task_arn = match result {
-        Ok(resp) => resp["task_arn"].as_str().unwrap_or("").to_string(),
-        Err(e) => return handle_result(Err(e)).await.into_response(),
-    };
-
-    // Extract task ID from ARN: arn:aws:ecs:region:account:task/cluster/TASK_ID
-    let task_id = task_arn.split('/').last().unwrap_or(&task_arn).to_string();
-
-    log::info!("Task ARN: {}, Task ID: {}", task_arn, task_id);
-
-    // Insert deployment record with variables into database using task ID
-    // This allows the runner to query the deployment and get variables
-    if let Err(e) = insert_deployment_record(&payload, &variables, &task_id).await {
-        log::error!("Failed to insert deployment record: {}", e);
-        return handle_result(Err(e)).await.into_response();
-    }
-
-    handle_result(Ok(json!({
-        "task_arn": task_arn,
-        "job_id": task_id
-    })))
-    .await
-    .into_response()
-}
-
-async fn insert_deployment_record(
-    payload: &env_defs::ApiInfraPayload,
-    variables: &serde_json::Value,
-    job_id: &str,
-) -> Result<(), anyhow::Error> {
-    use env_common::interface::GenericCloudHandler;
-
-    let handler = GenericCloudHandler::workload(&payload.project_id, &payload.region).await;
-
-    let payload_with_variables = env_defs::ApiInfraPayloadWithVariables {
-        payload: payload.clone(),
-        variables: variables.clone(),
-    };
-
-    env_common::insert_request_event(&handler, &payload_with_variables, job_id).await
-}
-
 async fn get_job_status_http(
     Path((project, region, rest)): Path<(String, String, String)>,
 ) -> impl IntoResponse {
@@ -1764,6 +1072,7 @@ async fn get_meta_info() -> impl IntoResponse {
 mod tests {
     use super::*;
     use axum::body::Body;
+    use axum::http::Request;
     use tower::ServiceExt;
 
     #[tokio::test]
