+++ internal-api/src/lib.rs
@@ -5,11 +5,15 @@ pub mod aws_handlers;
 #[cfg(feature = "azure")]
 pub mod azure_handlers;
 mod common;
+mod deployment_routes;
 pub mod handlers;
+mod http_authz;
+mod http_response;
 pub mod http_router;
 #[cfg(feature = "local")]
 pub mod local_setup;
 mod publish_auth;
+mod publish_routes;
 mod queries;
 
 pub use common::CloudRuntime;
